+++ src/dra/d_10798.c
@@ -20,7 +20,7 @@ s32 D_800B083C[] = {FIX(0), FIX(-0.75), FIX(0.75)};
 // tetra spirit out of bounds problem, add a value at beginning and end
 #ifdef VERSION_PC
 u8 D_800B0848[] = {0x00, 0x80, 0x80, 0x80, 0x90, 0x80, 0xA0, 0x80, 0xB0,
-                   0xA0, 0x80, 0xA0, 0x90, 0xA0, 0xA0, 0xA0, 0xB0, 0x03};
+                   0xA0, 0x80, 0xA0, 0x90, 0xA0, 0xA0, 0xA0, 0xB0, 0xC0};
 #else
 u8 D_800B0848[] = {0x80, 0x80, 0x80, 0x90, 0x80, 0xA0, 0x80, 0xB0,
                    0xA0, 0x80, 0xA0, 0x90, 0xA0, 0xA0, 0xA0, 0xB0};
